refactor: header suspense 제거 및 Loadable 처리
장바구니 데이터 fetch와 관련없이 Header 컴포넌트는 랜더링 되어있도록 처리
